clone: fix cast of pidfd pointer for 32-bit platforms
Casting this to 64-bit in 32-bit mode causes the upper 32-bits to be
sign-extended for some reason. This causes an EINVAL when calling
clone3(). Use cast to uintptr_t instead to avoid this effect.

+++ include/proc/clone.hxx
@@ -77,7 +77,7 @@ struct COSMOS_API CloneArgs :
 	}
 
 	void setPidFD(PidFD &fd) {
-		this->pidfd = reinterpret_cast<uint64_t>(&fd.m_fd);
+		this->pidfd = reinterpret_cast<uintptr_t>(&fd.m_fd);
 	}
 
 	void setChildTID(ProcessID *pid) {
